+++ TODO
@@ -1,0 +1,4 @@
+캐릭터 나누기
+텍스트 나누기
+sfx 정리하기
+assets 불려오는 거 정리하기
